refactor: small code improvements
