+++ src/app.ts
@@ -1,6 +1,6 @@
 import isMulti from './helpers/isMulti';
-import getServer from './getServer';
-import getMasterServer from './getMasterServer';
+import getServer from './servers/getServer';
+import getMasterServer from './servers/getMasterServer';
 import cluster from 'cluster';
 import LoadBalancer from './utils/loadBalancer';
 import dotenv from 'dotenv';
